+++ src/Game/Init.hs
@@ -53,7 +53,8 @@ import           Game.Types
   , Step(..)
   , Inbox(..)
   , SoundBank(..)
-  , Player'SFX'Key(..) )
+  , Player'SFX'Key(..)
+  , CollisionModule(..) )
 import           Game.Jump (floating)
 import           Game.Sprite (loadSpriteSheet)
 
@@ -99,6 +100,7 @@ initSystems renderer = void $ do
       , Velocity $ V2 0 0
       -- , Acceleration $ V2 0 0
       , BoundingBox $ V2 1 1.55
+      , CollisionModule
       , Gravity
         { ascent  = initialJumpG
         , descent = initialFallG }
@@ -109,7 +111,6 @@ initSystems renderer = void $ do
           , flowLimit   = 50
           , counter     = 0  }
       , FlowEffectEmitter NotEmittingFlowEffect )
-    , Inbox []
     , SpriteSheet playerSpriteSheet (Animate.initPosition PlayerKey'RIdle) )
 
   newEntity ( -- camera
@@ -125,24 +126,29 @@ initSystems renderer = void $ do
   newEntity ( -- floor
       Position $ V2 0 (screenHeight - 1)
     -- , Friction floorFriction
+    , CollisionModule
     , BoundingBox (V2 screenWidth 1) )
 
   newEntity ( --floating platform
       Position $ V2 5 (screenHeight / 2)
     -- , Friction floorFriction
+    , CollisionModule
     , BoundingBox $ V2 6 1 )
 
   newEntity ( --floating platform 2
       Position $ V2 13 (screenHeight / 2)
     -- , Friction floorFriction
+    , CollisionModule
     , BoundingBox $ V2 6 1 )
 
   newEntity ( --wall
       Position $ V2 (screenWidth - 10) (screenHeight - 5)
+    , CollisionModule
     , BoundingBox $ V2 1 4 )
 
   newEntity ( --wall2
       Position $ V2 (screenWidth - 1) (screenHeight - 8)
+    , CollisionModule
     , BoundingBox $ V2 1 7 )
 
   newEntity ( --hard flow1
@@ -151,9 +157,9 @@ initSystems renderer = void $ do
         { ascent  = initialJumpG
         , descent = initialFallG }
     , Position $ V2 2 (screenHeight - 1.5)
+    , CollisionModule
     , BoundingBox $ V2 0.25 0.25
-    , Velocity $ V2 0 0
-    , Inbox [] )
+    , Velocity $ V2 0 0 )
 
   newEntity ( -- audio player
       SoundBank {
