+++ src/Game/Physics.hs
@@ -24,17 +24,7 @@ import           Apecs
   , modify )
 
 import           Game.World (System')
-import           Game.AABB
-  ( aabbCheck
-  , sweepAABB
-  , penetrationVector
-  , broadPhaseAABB
-  , inNarrowPhase )
-import           Game.Collision
-  ( toVector
-  , resolveBaseCollision
-  , resolveNormalVelocity
-  , inverseNormal )
+import           Game.Collision (handleCollisions)
 import           Game.Constants
   ( dT
   , frameDeltaSeconds
@@ -49,20 +39,9 @@ import           Game.Constants
 import           Game.Types
   ( Unit(..)
   , Jump(..), buttonPressed, isJumping, isGrounded
-  , Acceleration(..)
   , Position(..)
   , Velocity(..)
-  , Friction(..)
-  , BoundingBox(..)
-  , Gravity(..)
-  , PlayerInput(..)
-  , AABB(..), center, dims
-  , BoxEntity(..)
-  , Collision(..)
-  , CollisionNormal(..)
-  , PenetrationVector(..)
-  , CollisionTime(..)
-  , Inbox(..) )
+  , Gravity(..) )
 import Game.Jump
   ( landed
   , onGround
@@ -71,83 +50,8 @@ import Game.Jump
   , floating
   , falling )
 
-
 oneBumpPerSecond :: Unit
 oneBumpPerSecond = (4 / 32) * Unit frameDeltaSeconds
-
-
-hasVelComponent :: BoxEntity -> System' Bool
-hasVelComponent (_, _, e) = do
-  hasVelocity <- exists e (proxy :: Velocity)
-  return hasVelocity
-
-handleBaseCollision :: Entity -> Collision -> System' ()
-handleBaseCollision e c@(Collision _ _ _ _) = do
-  pv <- get e :: System' (Position, Velocity)
-  set e $ resolveBaseCollision c pv
-
-handleJumpCheck :: Entity -> Collision -> System' ()
-handleJumpCheck e (Collision _ normal _ _) = do
-  hasJump <- exists e (proxy :: Jump)
-  when (hasJump && normal == BottomNormal) $ do
-    jumpState <- get e :: System' Jump
-    when (jumpState == jumping) $ set e landed
-    when (jumpState == floating || jumpState == falling) $ set e onGround
-
-
-handlePostCollision :: Entity -> Collision -> System' ()
-handlePostCollision e c = do
-  -- handleFloor e c
-  handleJumpCheck e c
-
-type GetVelocity  = System' (Velocity)
-type GetAABB      = System' (BoundingBox, Position)
-type GetSweptAABB = System' (BoundingBox, Position, Velocity)
-
-
-testCollision :: Entity -> BoxEntity -> System' ()
-testCollision e (_, _, e') = do
-  -- get current collision information
-  (BoundingBox bb1, Position p1, v@(Velocity v')) <- get e :: GetSweptAABB
-  (BoundingBox bb2, Position p2) <- get e' :: GetAABB
-  let box1 = AABB { center = p1, dims = bb1 }
-      box2 = AABB { center = p2, dims = bb2 }
-      (collisionTime, normal) = sweepAABB v box1 box2
-      (pVec, pNormal) = penetrationVector box1 box2
-      pVector = PenetrationVector $ pVec * (toVector pNormal)
-      collision = Collision collisionTime normal pVector e'
-      lowCollisionTime = ((coerce collisionTime :: Double) * frameDeltaSeconds) < 0.00005
-      noPenetration = (abs <$> (coerce pVector :: V2 Unit)) == V2 0 0
-      hasZeroNormal = normal == NoneNormal
-      useSimpleResolution =
-            hasZeroNormal
-        ||  lowCollisionTime
-        || (hasZeroNormal && noPenetration)
-
-  if (useSimpleResolution)
-  -- collision too slow to use swept resolution, or we hit a corner
-  then do
-    let Velocity v'' = resolveNormalVelocity v pVector normal
-        willNotEscape = aabbCheck
-          (broadPhaseAABB (BoundingBox bb1) (Position p1) (Velocity v''))
-          box2
-        -- handle floor corners
-        p' = if (hasZeroNormal && not noPenetration && willNotEscape)
-             -- push out of wall
-             then p1 + (coerce pVector :: V2 Unit)
-             -- otherwise update normally
-             else p1 + (v'' ^* Unit frameDeltaSeconds)
-    set e (Velocity v'', Position p')
-    -- dispatchToInbox
-    --   (Collision collisionTime pNormal pVector e') e
-    -- let reverseVector = PenetrationVector $ negate <$> (coerce pVector :: V2 Unit)
-    --     reverseNormal = inverseNormal pNormal
-    -- dispatchToInbox
-    --   (Collision collisionTime reverseNormal reverseVector e) e'
-  -- we need a swept collision resolution
-  else handleBaseCollision e collision
-  -- move on to continued resolution
-  handlePostCollision e collision
 
 stepPosition :: (Velocity, Position) -> (Velocity, Position)
 stepPosition (v@(Velocity v'), Position p) =
@@ -161,30 +65,6 @@ handleNotOnGround entity = do
   when hasJump $ do
     jumpState <- get entity :: System' Jump
     when (jumpState == onGround || jumpState == landed) $ set entity falling
-
-handleCollisions :: System' ()
-handleCollisions = do
-  -- get all entities with a bounding box
-  allBoundingBoxes <- getAll :: System' [BoxEntity]
-
-  -- partition into moving boxes & static boxes
-  (dynamics, _) <- partitionM hasVelComponent allBoundingBoxes
-
-  mapM_ (\(bb@(BoundingBox bb'), p@(Position p'), entity) -> do
-       v@(Velocity v') <- get entity :: System' Velocity
-       let sweptBox = broadPhaseAABB bb p v
-           actives  = filter (inNarrowPhase entity sweptBox) allBoundingBoxes
-
-       -- run all collision updates
-       mapM_ (testCollision entity) actives
-
-       -- when no collisions happened, update position and velocity normally
-       when (length actives == 0) $ do
-         handleNotOnGround entity
-         -- update position based on time and velocity
-         let p'' = p' + (v' ^* Unit frameDeltaSeconds)
-         set entity (Position p'')
-     ) dynamics
 
 handleGravity :: (Gravity, Velocity) -> Velocity
 handleGravity (g@(Gravity _ _), Velocity (V2 vx vy)) =
