+++ src/Game/Util/AABB.hs
@@ -1,4 +1,4 @@
-module Game.AABB where
+module Game.Util.AABB where
 
 import Linear (V2(..))
 import Apecs (Entity)
@@ -13,9 +13,9 @@ import Game.Types
   , PenetrationVector(..)
   , Collision(..)
   , AABB(..)
-  , BoxEntity(..) )
+  , BoxEntity)
 import Game.Constants (frameDeltaSeconds)
-import Game.Collision (toVector)
+import Game.Util.Collision (toVector)
 
 aabbCheck :: AABB -> AABB -> Bool
 aabbCheck (AABB (V2 x1 y1) (V2 w1 h1)) (AABB (V2 x2 y2) (V2 w2 h2)) =
@@ -112,6 +112,6 @@ sweepAABB (Velocity (V2 (Unit b1vx) (Unit b1vy)))
              , yInvExit  / (b1vy * frameDeltaSeconds) )
 
 inNarrowPhase :: Entity -> AABB -> BoxEntity -> Bool
-inNarrowPhase e sweptBox (BoundingBox bb, Position p, e') =
+inNarrowPhase e sweptBox (_, BoundingBox bb, Position p, e') =
   (not $ e' == e) && (aabbCheck sweptBox $ AABB { center = p, dims = bb })
